Refactor setup functions into a separate module
Refactored functions for setting up the camera and renderer into a new module named setup.js for better code organization and separation of concerns. This change makes these setup functions more reusable and the main.js file more succinct and easier to read and maintain.
